New user validation working

+++ src/reviews/reviews.service.ts
@@ -1,5 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { Hospital } from 'src/hospitals/entities/hospital.entity';
+import { User } from 'src/users/entities/user.entity';
 import { Repository } from 'typeorm';
 import { CreateReviewDto } from './dto/create-review.dto';
 import { UpdateReviewDto } from './dto/update-review.dto';
@@ -26,24 +32,51 @@ export class ReviewsService {
     return review;
   }
 
-  create(createReviewDto: CreateReviewDto) {
-    const review = this.reviewRepository.create(createReviewDto);
-    return this.reviewRepository.save(review);
+  public async findReviewsByHospital(hospital) {
+    return await this.reviewRepository.find({
+      where: { hospital: hospital },
+    });
   }
 
-  async update(id: string, updateReviewDto: UpdateReviewDto) {
-    const review = await this.reviewRepository.preload({
-      id: +id,
-      ...updateReviewDto,
+  public async createReview(
+    createReviewDto: CreateReviewDto,
+    user: User,
+    hospital: Hospital,
+  ): Promise<Review> {
+    return await this.reviewRepository.save({
+      ...createReviewDto,
+      user,
+      hospital,
     });
+  }
+
+  public async update(
+    id: string,
+    updateReviewDto: UpdateReviewDto,
+    user: User,
+  ) {
+    const review = await this.findOne(id);
     if (!review) {
       throw new NotFoundException(`Review #${id} not found.`);
     }
-    return this.reviewRepository.save(review);
+    if (review.user.id !== user.id && user.isadmin === false) {
+      throw new ForbiddenException(
+        "You don't have permission to edit this review.",
+      );
+    }
+    return await this.reviewRepository.save({
+      ...review,
+      ...updateReviewDto,
+    });
   }
 
-  async remove(id: string) {
+  async remove(id: string, user: User) {
     const review = await this.findOne(id);
+    if (review.user.id !== user.id && user.isadmin === false) {
+      throw new ForbiddenException(
+        "You don't have permission to delete this review.",
+      );
+    }
     return this.reviewRepository.remove(review);
   }
 }
